--- a/Domain/RDD.Domain.Tests/CollectionMethodsTests.cs
+++ b/Domain/RDD.Domain.Tests/CollectionMethodsTests.cs
@@ -1,6 +1,7 @@
 ﻿using Moq;
 using Newtonsoft.Json;
 using Rdd.Domain.Exceptions;
+using Rdd.Domain.Helpers.Reflection;
 using Rdd.Domain.Json;
 using Rdd.Domain.Mocks;
 using Rdd.Domain.Models;
@@ -125,7 +126,7 @@ namespace Rdd.Domain.Tests
         class OverrideObjectPatcher<T> : ObjectPatcher<T>
             where T : class, new()
         {
-            public OverrideObjectPatcher(IPatcherProvider provider) : base(provider)
+            public OverrideObjectPatcher(IPatcherProvider provider) : base(provider, new ReflectionHelper())
             {
             }
 
@@ -178,7 +179,7 @@ namespace Rdd.Domain.Tests
             {
                 var repo = new Repository<Hierarchy>(_fixture.InMemoryStorage, new Mock<IRightExpressionsHelper<Hierarchy>>().Object);
                 var instanciator = new BaseClassInstanciator<Hierarchy>(new InheritanceConfiguration());
-                var collection = new RestCollection<Hierarchy, int>(repo, new ObjectPatcher<Hierarchy>(_fixture.PatcherProvider), instanciator);
+                var collection = new RestCollection<Hierarchy, int>(repo, new ObjectPatcher<Hierarchy>(_fixture.PatcherProvider, _fixture.ReflectionHelper), instanciator);
 
                 JsonConvert.DefaultSettings = () => new JsonSerializerSettings
                 {
@@ -197,7 +198,7 @@ namespace Rdd.Domain.Tests
         {
             var repo = new Repository<Hierarchy>(_fixture.InMemoryStorage, new Mock<IRightExpressionsHelper<Hierarchy>>().Object);
             var instanciator = new BaseClassInstanciator<Hierarchy>(new InheritanceConfiguration());
-            var collection = new RestCollection<Hierarchy, int>(repo, new BaseClassPatcher<Hierarchy>(_fixture.PatcherProvider, new InheritanceConfiguration()), instanciator);
+            var collection = new RestCollection<Hierarchy, int>(repo, new BaseClassPatcher<Hierarchy>(_fixture.PatcherProvider, _fixture.ReflectionHelper, new InheritanceConfiguration()), instanciator);
 
             JsonConvert.DefaultSettings = () => new JsonSerializerSettings
             {