--- a/Domain/RDD.Domain.Tests/Models/UsersCollection.cs
+++ b/Domain/RDD.Domain.Tests/Models/UsersCollection.cs
@@ -1,4 +1,5 @@
-﻿using Rdd.Domain.Models;
+﻿using Rdd.Domain.Helpers.Reflection;
+using Rdd.Domain.Models;
 using Rdd.Domain.Patchers;
 using System;
 
@@ -7,6 +8,6 @@ namespace Rdd.Domain.Tests.Models
     public class UsersCollection : RestCollection<User, Guid>
     {
         public UsersCollection(IRepository<User> repository, IPatcherProvider patcherProvider, IInstanciator<User> instanciator)
-            : base(repository, new ObjectPatcher<User>(patcherProvider), instanciator) { }
+            : base(repository, new ObjectPatcher<User>(patcherProvider, new ReflectionHelper()), instanciator) { }
     }
 }