--- a/Domain/RDD.Domain.Tests/Models/UsersCollectionWithParameters.cs
+++ b/Domain/RDD.Domain.Tests/Models/UsersCollectionWithParameters.cs
@@ -1,4 +1,5 @@
-﻿using Rdd.Domain.Models;
+﻿using Rdd.Domain.Helpers.Reflection;
+using Rdd.Domain.Models;
 using Rdd.Domain.Patchers;
 
 namespace Rdd.Domain.Tests.Models
@@ -6,6 +7,6 @@ namespace Rdd.Domain.Tests.Models
     public class UsersCollectionWithParameters : RestCollection<UserWithParameters, int>
     {
         public UsersCollectionWithParameters(IRepository<UserWithParameters> repository, IPatcherProvider patcherProvider, IInstanciator<UserWithParameters> instanciator)
-            : base(repository, new ObjectPatcher<UserWithParameters>(patcherProvider), instanciator) { }
+            : base(repository, new ObjectPatcher<UserWithParameters>(patcherProvider, new ReflectionHelper()), instanciator) { }
     }
 }