--- a/Domain/RDD.Domain.Tests/PatchersTests.cs
+++ b/Domain/RDD.Domain.Tests/PatchersTests.cs
@@ -1,7 +1,11 @@
-﻿using Rdd.Domain.Exceptions;
+﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Rdd.Domain.Exceptions;
+using Rdd.Domain.Helpers.Reflection;
 using Rdd.Domain.Json;
 using Rdd.Domain.Patchers;
 using Rdd.Domain.Tests.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Xunit;
@@ -32,11 +36,23 @@ namespace Rdd.Domain.Tests
             public int[] ArrayOfInt { get; set; }
         }
 
-        IPatcherProvider PatcherProvider;
+        IServiceProvider ServiceProvider;
+        IPatcherProvider PatcherProvider => ServiceProvider.GetService<IPatcherProvider>();
+        IReflectionHelper ReflectionHelper => ServiceProvider.GetService<IReflectionHelper>();
 
         public PatchersTests()
         {
-            PatcherProvider = new PatcherProvider();
+            var services = new ServiceCollection();
+
+            services.TryAddSingleton<IReflectionHelper, ReflectionHelper>();
+            services.TryAddSingleton<IPatcherProvider, PatcherProvider>();
+            services.TryAddSingleton<EnumerablePatcher>();
+            services.TryAddSingleton<DictionaryPatcher>();
+            services.TryAddSingleton<ValuePatcher>();
+            services.TryAddSingleton<DynamicPatcher>();
+            services.TryAddSingleton<ObjectPatcher>();
+
+            ServiceProvider = services.BuildServiceProvider();
         }
 
         [Fact]
@@ -45,7 +61,7 @@ namespace Rdd.Domain.Tests
             var input = @"{intDico: { 1: 1, 2: 3 }, customFields: { 1: {code:""hey !""}, 2: {code:""salut""} }}";
             var json = new JsonParser().Parse(input);
             var newPatched = new ToPatch();
-            IPatcher patcher = new ObjectPatcher(PatcherProvider);
+            IPatcher patcher = new ObjectPatcher(PatcherProvider, ReflectionHelper);
             patcher.Patch(newPatched, json);
 
             Assert.Equal(1, newPatched.IntDico[1]);
@@ -60,7 +76,7 @@ namespace Rdd.Domain.Tests
             var input = @"{customFields: { 1: null, 2: {code:""test""} }}";
             var json = new JsonParser().Parse(input);
             var newPatched = new ToPatch();
-            IPatcher patcher = new ObjectPatcher(PatcherProvider);
+            IPatcher patcher = new ObjectPatcher(PatcherProvider, ReflectionHelper);
             patcher.Patch(newPatched, json);
 
             Assert.Null(newPatched.CustomFields[1]);
@@ -73,7 +89,7 @@ namespace Rdd.Domain.Tests
             var input = @"{keyStringValueStringDico: { ""yo"": null, ""man"": ""test"" }}";
             var json = new JsonParser().Parse(input);
             var newPatched = new ToPatch();
-            IPatcher patcher = new ObjectPatcher(PatcherProvider);
+            IPatcher patcher = new ObjectPatcher(PatcherProvider, ReflectionHelper);
             patcher.Patch(newPatched, json);
 
             Assert.Null(newPatched.KeyStringValueStringDico["yo"]);
@@ -86,7 +102,7 @@ namespace Rdd.Domain.Tests
             var input = @"{intNullableDico: { 1: null, 2: 3 }}";
             var json = new JsonParser().Parse(input);
             var newPatched = new ToPatch();
-            IPatcher patcher = new ObjectPatcher(PatcherProvider);
+            IPatcher patcher = new ObjectPatcher(PatcherProvider, ReflectionHelper);
             patcher.Patch(newPatched, json);
 
             Assert.Null(newPatched.IntNullableDico[1]);
@@ -99,7 +115,7 @@ namespace Rdd.Domain.Tests
             var input = @"{stringDico: { 1: null, 2: ""test"" }}";
             var json = new JsonParser().Parse(input);
             var newPatched = new ToPatch();
-            IPatcher patcher = new ObjectPatcher(PatcherProvider);
+            IPatcher patcher = new ObjectPatcher(PatcherProvider, ReflectionHelper);
             patcher.Patch(newPatched, json);
 
             Assert.Null(newPatched.StringDico[1]);
@@ -112,7 +128,7 @@ namespace Rdd.Domain.Tests
             Assert.Throws<BadRequestException>(() =>
             {
                 var json = new JsonParser().Parse(@"{intDico: { 1: null, 2: 3 }}");
-                IPatcher patcher = new ObjectPatcher(PatcherProvider);
+                IPatcher patcher = new ObjectPatcher(PatcherProvider, ReflectionHelper);
                 patcher.Patch(new ToPatch(), json);
             });
         }
@@ -123,7 +139,7 @@ namespace Rdd.Domain.Tests
             var input = @"{days: {code: ""Tue""}}";
             var json = new JsonParser().Parse(input);
             var newPatched = new ToPatch();
-            IPatcher patcher = new ObjectPatcher(PatcherProvider);
+            IPatcher patcher = new ObjectPatcher(PatcherProvider, ReflectionHelper);
             patcher.Patch(newPatched, json);
 
             Assert.Equal(TestEnum.Tue, newPatched.Days.Value);
@@ -135,7 +151,7 @@ namespace Rdd.Domain.Tests
             var input = @"{aCollection: [{mon: ""object""}, {autre: ""test""}]}";
             var json = new JsonParser().Parse(input);
             var newPatched = new ToPatch();
-            IPatcher patcher = new ObjectPatcher(PatcherProvider);
+            IPatcher patcher = new ObjectPatcher(PatcherProvider, ReflectionHelper);
             patcher.Patch(newPatched, json);
 
             Assert.Equal(2, newPatched.ACollection.Count);
@@ -149,7 +165,7 @@ namespace Rdd.Domain.Tests
             var input = @"{arrayOfInt: [1, 2, 3, 4]}";
             var json = new JsonParser().Parse(input);
             var newPatched = new ToPatch();
-            IPatcher patcher = new ObjectPatcher(PatcherProvider);
+            IPatcher patcher = new ObjectPatcher(PatcherProvider, ReflectionHelper);
             patcher.Patch(newPatched, json);
 
             Assert.Equal(4, newPatched.ArrayOfInt.Length);
@@ -163,7 +179,7 @@ namespace Rdd.Domain.Tests
         public void PatchFromAnonymous()
         {
             var newPatched = new ToPatch();
-            IPatcher patcher = new ObjectPatcher(PatcherProvider);
+            IPatcher patcher = new ObjectPatcher(PatcherProvider, ReflectionHelper);
             patcher.PatchFromAnonymous(newPatched, new { arrayOfInt = new[] { 1, 2, 3, 4 } });
 
             Assert.Equal(4, newPatched.ArrayOfInt.Length);