--- a/Domain/RDD.Domain/Helpers/Reflection/IReflectionHelper.cs
+++ b/Domain/RDD.Domain/Helpers/Reflection/IReflectionHelper.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Reflection;
+
+namespace Rdd.Domain.Helpers.Reflection
+{
+    public interface IReflectionHelper
+    {
+        IReadOnlyCollection<PropertyInfo> GetProperties(Type type);
+
+        object GetValue(object target, PropertyInfo property);
+        void SetValue(object target, PropertyInfo property, object value);
+
+        bool IsPseudoValue(Type type);
+    }
+}