--- a/Domain/RDD.Domain/RDD.Domain.csproj
+++ b/Domain/RDD.Domain/RDD.Domain.csproj
@@ -6,6 +6,7 @@
   </PropertyGroup>
   
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.1.1" />
     <PackageReference Include="Microsoft.Extensions.Primitives" Version="2.1.1" />
     <PackageReference Include="NExtends" Version="2.2.2" />
   </ItemGroup>