--- a/Web/RDD.Web.Tests/Models/Department.cs
+++ b/Web/RDD.Web.Tests/Models/Department.cs
@@ -3,11 +3,14 @@ using System.Collections.Generic;
 
 namespace Rdd.Web.Tests.Models
 {
+    public enum Test { A = 0, B = 10 }
     public class Department : EntityBase<Department, int>
     {
         public override int Id { get; set; }
         public override string Name { get; set; }
         public ICollection<User> Users { get; set; }
+
+        public Test? Enum { get; set; }
 
         public Department()
         {