… Benchmarks/Rdd.Benchmarks/Rdd.Benchmarks.sln
… Benchmarks/Rdd.Benchmarks/Rdd.Benchmarks/Program.cs
… Benchmarks/Rdd.Benchmarks/Rdd.Benchmarks/Rdd.Benchmarks.csproj
… Benchmarks/Rdd.Benchmarks/Rdd.Benchmarks/RddVsNewtonsoft.cs
… Domain/RDD.Domain/Patchers/PatcherProvider.cs
… Web/RDD.Web.Tests/Serialization/FieldsTests.cs
… Web/RDD.Web.Tests/Serialization/PropertySerializerTests.cs
… Web/RDD.Web/Helpers/RddServiceCollectionExtensions.cs
… Web/RDD.Web/Serialization/Providers/ISerializerProvider.cs
… Web/RDD.Web/Serialization/Providers/SerializerProvider.cs
… Web/RDD.Web/Serialization/Reflection/IReflectionProvider.cs
… Web/RDD.Web/Serialization/Reflection/ReflectionProvider.cs
… Web/RDD.Web/Serialization/Serializers/ArraySerializer.cs
… Web/RDD.Web/Serialization/Serializers/BaseClassSerializer.cs
… Web/RDD.Web/Serialization/Serializers/CultureInfoSerializer.cs
… Web/RDD.Web/Serialization/Serializers/DateTimeSerializer.cs
… Web/RDD.Web/Serialization/Serializers/DictionarySerializer.cs
… Web/RDD.Web/Serialization/Serializers/EntitySerializer.cs
… Web/RDD.Web/Serialization/Serializers/FuncSerializer.cs
… Web/RDD.Web/Serialization/Serializers/MetadataSerializer.cs
… Web/RDD.Web/Serialization/Serializers/ObjectSerializer.cs
… Web/RDD.Web/Serialization/Serializers/SelectionSerializer.cs
… Web/RDD.Web/Serialization/Serializers/ValueSerializer.cs
… docs/CHANGELOG.md