Ce week-end, je me suis mis en tête de regarder l'état des performances de la serialization RDD, notamment par rapport à celles de Newtonsoft.
J'ai fait mes tests sur une simple collection de 100 petits objets, contenant une propriété de 100 petits éléments.
Je me suis rendu compte qu'en l'absence de fields, Rdd était 20 fois plus lent que le natif, et allouait 80X plus de mémoire.
En présence de fields décrivant exactement ce qui était désiré, Rdd était environ 7 fois plus lent, et allouait 40X plus de mémoire.

J'ai donc travaillé sur l'algo, et désormais:
 - la différence entre avec et sans fields n'est plus significative
 - Rdd est environ 1.4 plus lent, et alloue 8X plus de mémoire.

cf résultats plus bas.
Détails sous les résultats


### No fields

**Before**

Method |       Mean |     Error |    StdDev | Scaled | ScaledSD |      Gen 0 |   Allocated |
----------- |-----------:|----------:|----------:|-------:|---------:|-----------:|------------:|
Rdd | 155.549 ms | 3.0319 ms | 5.1484 ms |  21.66 |     0.86 | 13250.0000 | 40723.38 KB |
Newtonsoft |   7.186 ms | 0.1360 ms | 0.1566 ms |   1.00 |     0.00 |    78.1250 |   245.93 KB |
 
**After**

Method |      Mean |     Error |    StdDev | Scaled |    Gen 0 |  Allocated |
----------- |----------:|----------:|----------:|-------:|---------:|-----------:|
Rdd | 10.566 ms | 0.0813 ms | 0.0761 ms |   1.45 | 656.2500 | 2022.32 KB |
Newtonsoft |  7.272 ms | 0.0262 ms | 0.0245 ms |   1.00 |  78.1250 |  245.93 KB |
 
 
### With explicit fields

**Before**

Method |      Mean |     Error |    StdDev |    Median | Scaled | ScaledSD |     Gen 0 |  Allocated |
----------- |----------:|----------:|----------:|----------:|-------:|---------:|----------:|-----------:|
Rdd | 48.670 ms | 0.9559 ms | 1.9953 ms | 49.058 ms |   6.94 |     0.40 | 2909.0909 | 9121.62 KB |
 Newtonsoft |  7.024 ms | 0.1398 ms | 0.2727 ms |  7.138 ms |   1.00 |     0.00 |   78.1250 |  245.93 KB |

**After**

Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |    Gen 0 |  Allocated |
----------- |---------:|----------:|----------:|-------:|---------:|---------:|-----------:|
Rdd | 9.970 ms | 0.1977 ms | 0.2571 ms |   1.34 |     0.04 | 656.2500 | 2022.32 KB |
Newtonsoft | 7.464 ms | 0.1233 ms | 0.1153 ms |   1.00 |     0.00 |  78.1250 |  245.93 KB |

### Améliorations
Voici la liste des différentes améliorations apportées pour remonter les performances de la serialization Rdd:
 - mis en cache des fields par défaut (environ 70ms !)
 - les serializers sont enregistrés en singleton au lieu d'être instanciés à chaque fois via activator (environ 60ms ! )
 - le serializer met en cache par type le serializer à renvoyer (environ 11ms)
 - Utilisation de Count != 0 au lieu de !.Any() (impact environ 2ms)
 - exposer le ValueProvider au proche du PropertyExpression au lieu du reflectionHelper (environ 1.5ms)
 - utilisation de dico au lieu de concurrentdico (pas franchement utilise ici) (environ 1ms)

Sans changer dramatiquement l'archi du système, je dois dire que je ne vois pas vraiment d'amélioration pour le moment. En l'état, la sérialization est proche des standards de performances du web